Update
Criação da rota de cadastro de cliente, criação da tela de loading final, melhorias no TimeTextPicker, melhorias no goBack dá página ConfigDay e ConfigInterval
