+++ src/src/routes/RegisterRoute.jsx
@@ -13,7 +13,8 @@ import { Interval } from '../pages/ConfigInterval';
 import { DisplacementFee } from '../pages/DisplacementFee';
 import { Services } from '../pages/ConfigServices';
 import { ServiceDetails } from '../pages/ServiceDetails';
-import { TestRegister } from '../pages/newRegisterTest';
+import { Register } from '../pages/Register';
+import { HomePage } from '../pages/HomePage';
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
 
@@ -24,7 +25,7 @@ export default function RegisterRoute() {
   const Stack = createNativeStackNavigator();
   return (
 
-    <Stack.Navigator initialRouteName='Services'>
+    <Stack.Navigator initialRouteName='Email'>
 
       <Stack.Screen name='Email' component={WhatsEmail} options={{
         title: 'Qual é o seu e-mail?',
@@ -123,6 +124,9 @@ export default function RegisterRoute() {
         headerTintColor: WhiteColor
       }} />
 
+      <Stack.Screen name='Register' component={Register} options={{ headerShown: false }} />
+
+      <Stack.Screen name='Home' component={HomePage} options={{ headerShown: false }} />
 
     </Stack.Navigator>
