+++ src/src/routes/index.jsx
@@ -19,7 +19,7 @@ const Stack = createNativeStackNavigator();
 
 export function Route() {
   return (
-    <Stack.Navigator initialRouteName='RegisterRoute'>
+    <Stack.Navigator initialRouteName='AppRoute'>
       <Stack.Screen name="AppRoute" component={AppRoute} options={{ headerShown: false }} />
       <Stack.Screen name="RegisterRoute" component={RegisterRoute} options={{ headerShown: false }} />
     </Stack.Navigator>
